Extracted RegenerateAutoReplayITReferences class (containing the main method of AutoReplayIT) into the tools project. Gradle depenencies still need to be adjusted.
